backend/app/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
sqlalchemy>=2.0.0
pydantic[email]>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.1
boto3>=1.26.0
httpx>=0.24.0
google-auth>=2.20.0
psycopg2-binary>=2.9.6
python-multipart>=0.0.9
pyswisseph>=2.10.3.1
requests>=2.31.0
alembic>=1.11.0
